requirements.txt
discord.py[voice] # main discord library
#git+https://github.com/Rapptz/discord.py

yt-dlp      # used to download audio
python-ffmpeg # used to apply audio filters and such
#spotdl      # used to download audio ( from spotify )

#mutagen     # used to read metadata from audio files

psutil      # gets system info, used for info command and such

pillow      # for image manipulation
pdf2image   # translates pdfs to pngs

beautifulsoup4 # for handling html data

python-dateutil # extension for the datetime library

#midiutil
#sound_to_midi
#librosa

requests